+++ gems/aws-sdk-s3/lib/aws-sdk-s3/customizations.rb
@@ -18,8 +18,11 @@ require 'aws-sdk-s3/presigner'
 
 # s3 express session auth
 require 'aws-sdk-s3/express_credentials'
-require 'aws-sdk-s3/express_credentials_cache'
 require 'aws-sdk-s3/express_credentials_provider'
+
+# s3 access grants auth
+require 'aws-sdk-s3/access_grants_credentials'
+require 'aws-sdk-s3/access_grants_credentials_provider'
 
 # customizations to generated classes
 require 'aws-sdk-s3/customizations/bucket'