+++ gems/aws-sdk-s3/spec/client/region_detection_spec.rb
@@ -33,9 +33,9 @@ module Aws
         'WlSrUk+8d2/rvcnEv2QXer0=</HostId></Error>'
       end
 
-      before(:each) do
+      before do
         allow($stderr).to receive(:write)
-        S3::BUCKET_REGIONS.clear
+        Aws::S3.bucket_region_cache.clear
       end
 
       context 'accessing us-west-2 bucket using classic endpoint' do
@@ -93,7 +93,7 @@ module Aws
         end
 
         it 'does not redirect custom endpoints when the region is cached' do
-          S3::BUCKET_REGIONS['bucket'] = 'us-west-2'
+          Aws::S3.bucket_region_cache['bucket'] = 'us-west-2'
           stub_request(:put, 'http://bucket.localhost:9000/key')
             .to_return(status: [200, 'Ok'])